Update Kafka services
